상세화면 조금 안맞는 부분 수정 / new 네이밍 제거
